+++ Demo/Setup.swift
@@ -5,20 +5,6 @@
 import FINNSetup
 import UIKit
 
-struct DemoVertical: Vertical {
-    let name: String
-    let title: String
-    var isCurrent: Bool
-    let isExternal: Bool
-
-    init(name: String, title: String, isCurrent: Bool = false, isExternal: Bool = false) {
-        self.name = name
-        self.title = title
-        self.isCurrent = isCurrent
-        self.isExternal = isExternal
-    }
-}
-
 class Setup {
     var filterContainer: FilterContainer
